+++ server/src/main/java/com/team_60/Mocco/alarm/service/AlarmService.java
@@ -15,4 +15,5 @@ public interface AlarmService {
     void createAlarmWhenStudyOpen(Study study);
     Alarm createAlarmWhenStudyNotOpen(Study study);
     Alarm createAlarmWhenProposalApprove(Study study, Member member);
+    public Alarm findVerifiedAlarm(long alarmId);
 }
