+++ server/src/main/java/com/team_60/Mocco/chatting/entity/Chatting.java
@@ -1,6 +1,6 @@
 package com.team_60.Mocco.chatting.entity;
 
-import com.team_60.Mocco.audit.Auditable;
+import com.team_60.Mocco.base_entity.BaseEntity;
 import com.team_60.Mocco.member.entity.Member;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,7 +12,7 @@ import javax.persistence.*;
 @Getter
 @Setter
 @Entity
-public class Chatting extends Auditable {
+public class Chatting extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long chattingId;
