+++ server/src/main/java/com/team_60/Mocco/helper/aop/AuthenticationServiceLocal.java
@@ -1,0 +1,20 @@
+package com.team_60.Mocco.helper.aop;
+
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.http.HttpServletRequest;
+
+@Component
+@Profile("!deploy")
+public class AuthenticationServiceLocal {
+
+    public void AuthenticationCheckWithId(String parameterName, long id) { // patch, delete (id값 받는 경우)
+    }
+
+    public void AuthenticationCheckWithDto(Object requestBody) { //post
+    }
+
+    public void AuthenticationCheckStudyMember(long studyId) { //studyRoom
+    }
+}
