+++ server/src/main/java/com/team_60/Mocco/member/controller/MemberSubController.java
@@ -2,7 +2,7 @@ package com.team_60.Mocco.member.controller;
 
 import com.team_60.Mocco.dto.SingleResponseDto;
 import com.team_60.Mocco.helper.aop.AuthenticationService;
-import com.team_60.Mocco.helper.aop.AuthenticationServiceDeploy;
+import com.team_60.Mocco.helper.aop.AuthenticationServiceLocal;
 import com.team_60.Mocco.helper.httpclient.GithubRestClient;
 import com.team_60.Mocco.helper.httpclient.dto.GithubRestClientDto;
 import com.team_60.Mocco.helper.upload.ImageUploadType;
@@ -27,7 +27,7 @@ public class MemberSubController {
     private final S3ImageUpload imageUpload;
     private final MemberMapper mapper;
     private final MemberService memberService;
-    private final AuthenticationService authenticationService;
+    private final AuthenticationServiceLocal authenticationService;
     private final GithubRestClient githubRestClient;
 
     @PostMapping("/image")
