+++ server/src/main/java/com/team_60/Mocco/member/entity/MyInfo.java
@@ -1,6 +1,6 @@
 package com.team_60.Mocco.member.entity;
 
-import com.team_60.Mocco.audit.Auditable;
+import com.team_60.Mocco.base_entity.BaseEntity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,7 +11,7 @@ import javax.persistence.*;
 @Getter
 @Setter
 @Entity
-public class MyInfo extends Auditable {
+public class MyInfo extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long myInfoId;
