+++ server/src/main/java/com/team_60/Mocco/proposal/controller/ProposalController.java
@@ -2,7 +2,7 @@ package com.team_60.Mocco.proposal.controller;
 
 import com.team_60.Mocco.dto.SingleResponseDto;
 import com.team_60.Mocco.helper.aop.AuthenticationService;
-import com.team_60.Mocco.helper.aop.AuthenticationServiceDeploy;
+import com.team_60.Mocco.helper.aop.AuthenticationServiceLocal;
 import com.team_60.Mocco.proposal.dto.ProposalDto;
 import com.team_60.Mocco.proposal.entity.Proposal;
 import com.team_60.Mocco.proposal.mapper.ProposalMapper;
@@ -22,7 +22,7 @@ public class ProposalController {
 
     private final ProposalService proposalService;
     private final ProposalMapper mapper;
-    private final AuthenticationService authenticationService;
+    private final AuthenticationServiceLocal authenticationService;
 
     @GetMapping
     public ResponseEntity getProposalsByStudyId(@RequestParam("study-id") long studyId){
@@ -35,8 +35,8 @@ public class ProposalController {
     }
 
     @PostMapping
-    public ResponseEntity postProposal(@RequestBody ProposalDto.Post requestBody, HttpServletRequest request){
-        authenticationService.AuthenticationCheckWithDto(requestBody,request);
+    public ResponseEntity postProposal(@RequestBody ProposalDto.Post requestBody){
+        authenticationService.AuthenticationCheckWithDto(requestBody);
         Proposal proposal = mapper.proposalPostDtoToProposal(requestBody);
         Proposal postProposal = proposalService.createProposal(proposal);
         ProposalDto.Response response = mapper.proposalToProposalResponseDto(postProposal);
