+++ server/src/main/java/com/team_60/Mocco/proposal/entity/Proposal.java
@@ -1,6 +1,6 @@
 package com.team_60.Mocco.proposal.entity;
 
-import com.team_60.Mocco.audit.Auditable;
+import com.team_60.Mocco.base_entity.BaseEntity;
 import com.team_60.Mocco.member.entity.Member;
 import com.team_60.Mocco.study.entity.Study;
 import lombok.Getter;
@@ -13,7 +13,7 @@ import javax.persistence.*;
 @Setter
 @NoArgsConstructor
 @Entity
-public class Proposal extends Auditable {
+public class Proposal extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long proposalId;
